requirements.txt
asgiref==3.5.2
Django==4.1.1
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
Markdown==3.4.1
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2
